+++ src/OpenTelemetry/Metrics/Exemplar/AlignedHistogramBucketExemplarReservoir.cs
@@ -8,92 +8,28 @@ namespace OpenTelemetry.Metrics;
 /// <summary>
 /// The AlignedHistogramBucketExemplarReservoir implementation.
 /// </summary>
-internal sealed class AlignedHistogramBucketExemplarReservoir : ExemplarReservoir
+internal sealed class AlignedHistogramBucketExemplarReservoir : FixedSizeExemplarReservoir
 {
-    private readonly Exemplar[] runningExemplars;
-    private readonly Exemplar[] tempExemplars;
-
-    public AlignedHistogramBucketExemplarReservoir(int length)
+    public AlignedHistogramBucketExemplarReservoir(int numberOfBuckets)
+        : base(numberOfBuckets + 1)
     {
-        this.runningExemplars = new Exemplar[length + 1];
-        this.tempExemplars = new Exemplar[length + 1];
     }
 
-    public override void Offer(long value, ReadOnlySpan<KeyValuePair<string, object?>> tags, int index = default)
+    public override void Offer(in ExemplarMeasurement<long> measurement)
     {
-        this.OfferAtBoundary(value, tags, index);
+        Debug.Assert(
+            measurement.ExplicitBucketHistogramBucketIndex != -1,
+            "ExplicitBucketHistogramBucketIndex was -1");
+
+        this.UpdateExemplar(measurement.ExplicitBucketHistogramBucketIndex, in measurement);
     }
 
-    public override void Offer(double value, ReadOnlySpan<KeyValuePair<string, object?>> tags, int index = default)
+    public override void Offer(in ExemplarMeasurement<double> measurement)
     {
-        this.OfferAtBoundary(value, tags, index);
-    }
+        Debug.Assert(
+            measurement.ExplicitBucketHistogramBucketIndex != -1,
+            "ExplicitBucketHistogramBucketIndex was -1");
 
-    public override Exemplar[] Collect(ReadOnlyTagCollection actualTags, bool reset)
-    {
-        for (int i = 0; i < this.runningExemplars.Length; i++)
-        {
-            this.tempExemplars[i] = this.runningExemplars[i];
-            if (this.runningExemplars[i].FilteredTags != null)
-            {
-                // TODO: Better data structure to avoid this Linq.
-                // This is doing filtered = alltags - storedtags.
-                // TODO: At this stage, this logic is done inside Reservoir.
-                // Kinda hard for end users who write own reservoirs.
-                // Evaluate if this logic can be moved elsewhere.
-                // TODO: The cost is paid irrespective of whether the
-                // Exporter supports Exemplar or not. One idea is to
-                // defer this until first exporter attempts read.
-                this.tempExemplars[i].FilteredTags = this.runningExemplars[i].FilteredTags!.Except(actualTags.KeyAndValues.ToList()).ToList();
-            }
-
-            if (reset)
-            {
-                this.runningExemplars[i].Timestamp = default;
-            }
-        }
-
-        return this.tempExemplars;
-    }
-
-    private void OfferAtBoundary(double value, ReadOnlySpan<KeyValuePair<string, object?>> tags, int index)
-    {
-        ref var exemplar = ref this.runningExemplars[index];
-        exemplar.Timestamp = DateTimeOffset.UtcNow;
-        exemplar.DoubleValue = value;
-        exemplar.TraceId = Activity.Current?.TraceId;
-        exemplar.SpanId = Activity.Current?.SpanId;
-
-        if (tags == default)
-        {
-            // default tag is used to indicate
-            // the special case where all tags provided at measurement
-            // recording time are stored.
-            // In this case, Exemplars does not have to store any tags.
-            // In other words, FilteredTags will be empty.
-            return;
-        }
-
-        if (exemplar.FilteredTags == null)
-        {
-            exemplar.FilteredTags = new List<KeyValuePair<string, object?>>(tags.Length);
-        }
-        else
-        {
-            // Keep the list, but clear contents.
-            exemplar.FilteredTags.Clear();
-        }
-
-        // Though only those tags that are filtered need to be
-        // stored, finding filtered list from the full tag list
-        // is expensive. So all the tags are stored in hot path (this).
-        // During snapshot, the filtered list is calculated.
-        // TODO: Evaluate alternative approaches based on perf.
-        // TODO: This is not user friendly to Reservoir authors
-        // and must be handled as transparently as feasible.
-        foreach (var tag in tags)
-        {
-            exemplar.FilteredTags.Add(tag);
-        }
+        this.UpdateExemplar(measurement.ExplicitBucketHistogramBucketIndex, in measurement);
     }
 }